Major: Updated Forum to read db.json
The forum pages now read their data from the db.json. The 'Post' method has not been added, so instead they read from the local storage when there is a change compared to the json file. Some changes were made to the .css files due to the need of a different font